Bug #329382 Pick-up-LOCAL_RELEASEKEY-for-x3542-for-signed-images
[bug number  ] 329382
[root cause  ] this will take a path to the key and cert used to sign and verify the fotas.
A typical path is "vendor/sprd/security/releasekey" and then vendor/sprd/security is expected to contain a .pk8 and a .x509.pem
for each of releasekey (so releasekey.pk8 and releasekey.x509.pem), testkey, shared, media and platform.
It is unacceptable to use the key at build/target/product/security/testkey since it is public.
[changes     ] local releasekey
[side effects] none
[self test   ] yes
[whether AOB ] no
[reviewers   ]

Change-Id: Ib25011c0316dcb61475fdfb73fb3da709d45c26b